server_py/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Cors==3.0.10
